# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AssessmentTaskCreated": {
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "status": {
            "enum": [
              "accepted",
              "queued"
            ],
            "title": "Status",
            "type": "string"
          },
          "task_contract": {
            "$ref": "#/components/schemas/AgentTaskContract"
          },
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "status",
          "task_contract"
        ],
        "title": "AssessmentTaskCreated",
        "type": "object"
      },
      "AssessmentTaskResult": {
        "properties": {
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          },
          "comments": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Comments",
            "type": "array"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvaluationArtifact"
              },
              {
                "type": "null"
              }
            ]
          },
          "plan_artifact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanArtifact"
              },
              {
                "type": "null"
              }
            ]
          },
          "report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssessmentReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "review_pending",
              "approved",
              "rejected",
              "escalated",
              "completed",
              "failed",
              "cancelled",
              "interrupted"
            ],
            "title": "Status",
            "type": "string"
          },
          "task_contract": {
            "$ref": "#/components/schemas/AgentTaskContract"
          },
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          },
          "version": {
            "default": 1,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "task_id",
          "status",
          "created_at",
          "task_contract"
        ],
        "title": "AssessmentTaskResult",
        "type": "object"
      },
      "Body_submit_assessment_api_v1_assessments_post": {
        "properties": {
          "collaborative_mode": {
            "default": true,
            "title": "Collaborative Mode",
            "type": "boolean"
          },
          "files": {
            "description": "Documents to assess",
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "phase": {
            "default": "auto",
            "enum": [
              "auto",
              "requirements",
              "design",
              "development",
              "testing",
              "deployment",
              "operations",
              "full_ssdlc"
            ],
            "title": "Phase",
            "type": "string"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "skill_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Id"
          }
        },
        "required": [
          "files"
        ],
        "title": "Body_submit_assessment_api_v1_assessments_post",
        "type": "object"
      },
...
      "CommentRequest": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "content"
        ],
        "title": "CommentRequest",
        "type": "object"
      },
...
      "EvaluationArtifact": {
        "additionalProperties": false,
        "properties": {
          "checks": {
            "items": {
              "$ref": "#/components/schemas/EvaluationCheck"
            },
            "minItems": 1,
            "title": "Checks",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evaluator": {
            "const": "deterministic_policy",
            "default": "deterministic_policy",
            "title": "Evaluator",
            "type": "string"
          },
          "outcome": {
            "enum": [
              "passed",
              "needs_review",
              "failed"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          },
          "version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "created_at",
          "outcome",
          "checks"
        ],
        "title": "EvaluationArtifact",
        "type": "object"
      },
      "EvaluationCheck": {
        "additionalProperties": false,
        "properties": {
          "details": {
            "title": "Details",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "passed": {
            "title": "Passed",
            "type": "boolean"
          },
          "required": {
            "default": true,
            "title": "Required",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "passed",
          "details"
        ],
        "title": "EvaluationCheck",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RemediationTracking": {
        "properties": {
          "due_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Refs",
            "type": "array"
          },
          "external_ticket": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Ticket"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "remediation_id": {
            "title": "Remediation Id",
            "type": "string"
          },
          "status": {
            "default": "open",
            "enum": [
              "open",
              "in_progress",
              "resolved",
              "verified",
              "closed"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "remediation_id"
        ],
        "title": "RemediationTracking",
        "type": "object"
      },
      "RemediationTrackingUpdateRequest": {
        "properties": {
          "due_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "evidence_refs": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Refs"
          },
          "external_ticket": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Ticket"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "open",
                  "in_progress",
                  "resolved",
                  "verified",
                  "closed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "RemediationTrackingUpdateRequest",
        "type": "object"
      },
...
      "ReviewActionRequest": {
        "properties": {
          "action": {
            "enum": [
              "approve",
              "reject",
              "comment",
              "escalate"
            ],
            "title": "Action",
            "type": "string"
          },
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "required": [
          "action"
        ],
        "title": "ReviewActionRequest",
        "type": "object"
      },
...
      "TrackedRemediation": {
        "properties": {
          "remediation": {
            "$ref": "#/components/schemas/Remediation"
          },
          "tracking": {
            "$ref": "#/components/schemas/RemediationTracking"
          }
        },
        "required": [
          "remediation",
          "tracking"
        ],
        "title": "TrackedRemediation",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Automated Security Assessment with LLMs & RAG",
    "title": "DocSentinel API",
    "version": "4.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/assessments": {
      "get": {
        "operationId": "list_assessments_api_v1_assessments_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "assignee",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assignee"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssessmentTaskResult"
                  },
                  "title": "Response List Assessments Api V1 Assessments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Assessments",
        "tags": [
          "assessments"
        ]
      },
      "post": {
        "description": "Submit an assessment task; returns task_id immediately for polling.",
        "operationId": "submit_assessment_api_v1_assessments_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_assessment_api_v1_assessments_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentTaskCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit Assessment",
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{task_id}": {
      "get": {
        "operationId": "get_assessment_result_api_v1_assessments__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentTaskResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Assessment Result",
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{task_id}/activity": {
      "get": {
        "operationId": "get_task_activity_api_v1_assessments__task_id__activity_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Task Activity",
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{task_id}/comments": {
      "post": {
        "operationId": "add_comment_api_v1_assessments__task_id__comments_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Comment",
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{task_id}/remediations": {
      "get": {
        "operationId": "list_tracked_remediations_api_v1_assessments__task_id__remediations_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TrackedRemediation"
                  },
                  "title": "Response List Tracked Remediations Api V1 Assessments  Task Id  Remediations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Tracked Remediations",
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{task_id}/remediations/{remediation_id}": {
      "post": {
        "operationId": "update_remediation_tracking_api_v1_assessments__task_id__remediations__remediation_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "remediation_id",
            "required": true,
            "schema": {
              "title": "Remediation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemediationTrackingUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemediationTracking"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Remediation Tracking",
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{task_id}/reuse": {
      "get": {
        "operationId": "get_reuse_candidates_api_v1_assessments__task_id__reuse_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "top_k",
            "required": false,
            "schema": {
              "default": 3,
              "title": "Top K",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Reuse Candidates",
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{task_id}/review": {
      "post": {
        "operationId": "review_assessment_api_v1_assessments__task_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Review Assessment",
        "tags": [
          "assessments"
        ]
      }
    },
...
  }
}